chore(listings): remove persistence mappers and inline admin mapping; re-export buildPagedArgs
