+++ packages/sthrift/graphql/src/schema/types/listing/item-listing.resolvers.ts
@@ -2,6 +2,7 @@ import type { GraphContext } from '../../../init/context.ts';
 // Domain types not needed in this file after refactor
 import { toGraphItem } from '../../../helpers/mapping.js';
 import type { CreateItemListingInput } from '../../builder/generated.js';
+import { buildPagedArgs } from '@sthrift/application-services';
 
 interface MyListingsArgs {
 	page: number;
@@ -11,70 +12,14 @@ interface MyListingsArgs {
 	sorter?: { field: string; order: 'ascend' | 'descend' };
 }
 
-// Small helpers extracted from resolver logic to keep resolvers focused on orchestration
-function buildPagedArgs(
-	args: MyListingsArgs,
-	opts?: { useDefaultStatuses?: boolean },
-) {
-	const { page, pageSize, searchText, statusFilters, sorter } = args;
-		let effectiveStatuses: string[] | undefined;
-		if (statusFilters && statusFilters.length > 0) {
-			effectiveStatuses = statusFilters;
-		} else if (opts?.useDefaultStatuses) {
-			effectiveStatuses = ['Appeal Requested', 'Blocked'];
-		} else {
-			effectiveStatuses = undefined;
-		}
-
-	const pagedArgs: Partial<MyListingsArgs> & { page: number; pageSize: number; sharerId?: string } = {
-		page,
-		pageSize,
-	};
-	if (searchText) pagedArgs.searchText = searchText;
-	if (effectiveStatuses !== undefined) pagedArgs.statusFilters = effectiveStatuses;
-	if (sorter) pagedArgs.sorter = sorter;
-	return pagedArgs;
-}
-
-function mapStateToStatus(state?: string) {
-	return state?.toString?.().trim?.() || 'Unknown';
-}
-
-function toAdminListing(listing: Record<string, unknown>) {
-	const getIso = (v: unknown) => {
-		try {
-			return (v as { toISOString?: () => string })?.toISOString?.() ?? '';
-		} catch {
-			return '';
-		}
-	};
-
-	// Small local shape so we can use dot-notation without `any`.
-	type ListingLike = {
-		sharingPeriodStart?: unknown;
-		sharingPeriodEnd?: unknown;
-		images?: string[];
-		createdAt?: unknown;
-		state?: string;
-		id?: string;
-		title?: string;
-	};
-	const l = listing as ListingLike;
-	const start = getIso(l.sharingPeriodStart);
-	const end = getIso(l.sharingPeriodEnd);
-	const images = l.images;
-	const createdAt = getIso(l.createdAt);
-	const state = l.state ?? undefined;
-	return {
-		id: (l.id as string) ?? '',
-		title: (l.title as string) ?? '',
-		image: images && images.length > 0 ? images[0] : null,
-		publishedAt: createdAt || null,
-		reservationPeriod: `${start.slice(0, 10)} - ${end.slice(0, 10)}`,
-		status: mapStateToStatus(state),
-		pendingRequestsCount: 0,
-	};
-}
+// Utility to safely convert date-like values to ISO strings
+const getIso = (v: unknown) => {
+	try {
+		return (v as { toISOString?: () => string })?.toISOString?.() ?? '';
+	} catch {
+		return '';
+	}
+};
 
 const itemListingResolvers = {
 	Query: {
@@ -91,10 +36,34 @@ const itemListingResolvers = {
 			const result = await context.applicationServices.Listing.ItemListing.queryPaged(pagedArgs);
 
 			return {
-				items: result.items.map(
-					//biome-ignore lint/suspicious/noExplicitAny: Mongoose document type is dynamic
-					(listing: any) => toAdminListing(listing),
-				),
+				items: result.items.map((listing) => {
+					// Map listing entity reference directly — mapping helpers removed per request
+					type ListingLike = {
+						sharingPeriodStart?: unknown;
+						sharingPeriodEnd?: unknown;
+						images?: string[];
+						createdAt?: unknown;
+						state?: string;
+						id?: string;
+						title?: string;
+					};
+					const l = listing as unknown as ListingLike;
+					// use shared getIso helper above
+					const start = getIso(l.sharingPeriodStart);
+					const end = getIso(l.sharingPeriodEnd);
+					const images = l.images;
+					const createdAt = getIso(l.createdAt);
+					const state = l.state ?? undefined;
+					return {
+						id: (l.id as string) ?? '',
+						title: (l.title as string) ?? '',
+						image: images && images.length > 0 ? images[0] : null,
+						publishedAt: createdAt || null,
+						reservationPeriod: `${start.slice(0, 10)} - ${end.slice(0, 10)}`,
+						status: state?.toString?.().trim?.() || 'Unknown',
+						pendingRequestsCount: 0,
+					};
+				}),
 				total: result.total,
 				page: result.page,
 				pageSize: result.pageSize,
@@ -106,7 +75,7 @@ const itemListingResolvers = {
 			args: { id: string },
 			context: GraphContext,
 		) => {
-			// TODO: SECURITY - Add admin role-based authorization check when admin role system is implemented
+			// Admin-note: role-based authorization should be implemented here (security)
 			const listing =
 				await context.applicationServices.Listing.ItemListing.queryById({
 					id: args.id,
@@ -122,15 +91,39 @@ const itemListingResolvers = {
 
 
 		adminListings: async (_parent: unknown, args: MyListingsArgs, context: GraphContext) => {
-			// TODO: SECURITY - Add admin role-based authorization check when admin role system is implemented
+			// Admin-note: role-based authorization should be implemented here (security)
 			const pagedArgs = buildPagedArgs(args, { useDefaultStatuses: true });
 			const result = await context.applicationServices.Listing.ItemListing.queryPaged(pagedArgs);
 
 			return {
-				items: result.items.map(
-					//biome-ignore lint/suspicious/noExplicitAny: Mongoose document type is dynamic
-					(listing: any) => toAdminListing(listing),
-				),
+				items: result.items.map((listing) => {
+					// Map listing entity reference directly — mapping helpers removed per request
+					type ListingLike = {
+						sharingPeriodStart?: unknown;
+						sharingPeriodEnd?: unknown;
+						images?: string[];
+						createdAt?: unknown;
+						state?: string;
+						id?: string;
+						title?: string;
+					};
+					const l = listing as unknown as ListingLike;
+					// use shared getIso helper above
+					const start = getIso(l.sharingPeriodStart);
+					const end = getIso(l.sharingPeriodEnd);
+					const images = l.images;
+					const createdAt = getIso(l.createdAt);
+					const state = l.state ?? undefined;
+					return {
+						id: (l.id as string) ?? '',
+						title: (l.title as string) ?? '',
+						image: images && images.length > 0 ? images[0] : null,
+						publishedAt: createdAt || null,
+						reservationPeriod: `${start.slice(0, 10)} - ${end.slice(0, 10)}`,
+						status: state?.toString?.().trim?.() || 'Unknown',
+						pendingRequestsCount: 0,
+					};
+				}),
 				total: result.total,
 				page: result.page,
 				pageSize: result.pageSize,
@@ -180,7 +173,7 @@ const itemListingResolvers = {
 			args: { id: string },
 			context: GraphContext,
 		) => {
-			// TODO: SECURITY - Add admin role-based authorization check when admin role system is implemented
+			// Admin-note: role-based authorization should be implemented here (security)
 			// Once implemented, use system-level permissions for admin operations
 			return await context.applicationServices.Listing.ItemListing.remove({ id: args.id });
 		},
@@ -190,7 +183,7 @@ const itemListingResolvers = {
 			args: { id: string },
 			context: GraphContext,
 		) => {
-			// TODO: SECURITY - Add admin role-based authorization check when admin role system is implemented
+			// Admin-note: role-based authorization should be implemented here (security)
 			// Once implemented, use system-level permissions for admin operations
 			return await context.applicationServices.Listing.ItemListing.unblock({ id: args.id });
 		},
